+++ packages/core/src/connectivity/scp-cf/destination/destination-cache.spec.ts
@@ -46,6 +46,7 @@ import {
 import { AuthenticationType, Destination } from './destination-service-types';
 import { getDestinationFromDestinationService } from './destination-from-service';
 import { parseDestination } from './destination';
+import { DestinationAuthToken } from '.';
 
 const destinationOne: Destination = {
   url: 'https://destination1.example',
@@ -98,13 +99,20 @@ function mockDestinationsWithSameName() {
   mockSubaccountDestinationsCall(nock, [dest], 200, providerServiceToken);
 }
 
-describe('caching destination integration tests', () => {
-  afterEach(() => {
+describe('destination cache', () => {
+  afterAll(() => {
     destinationCache.clear();
     destinationServiceCache.clear();
     nock.cleanAll();
   });
-  describe('test caching of retrieved entries', () => {
+
+  beforeEach(() => {
+    destinationCache.clear();
+    destinationServiceCache.clear();
+    nock.cleanAll();
+  });
+
+  describe('caching', () => {
     beforeEach(() => {
       mockVerifyJwt();
       mockServiceBindings();
@@ -294,7 +302,7 @@ describe('caching destination integration tests', () => {
   });
 
   describe('caching of destinations with special information (e.g. authTokens, certificates)', () => {
-    it('destinations with certificates are cached correctly', async () => {
+    it('destinations with certificates are cached ', async () => {
       mockServiceBindings();
       mockVerifyJwt();
       mockServiceToken();
@@ -535,12 +543,6 @@ describe('caching destination integration tests', () => {
       httpMocks.forEach(mock => expect(mock.isDone()).toBe(true));
     });
   });
-});
-
-describe('caching destination unit tests', () => {
-  beforeEach(() => {
-    destinationCache.clear();
-  });
 
   it('should cache the destination correctly', () => {
     const dummyJwt = { user_id: 'user', zid: 'tenant' };
@@ -659,6 +661,33 @@ describe('caching destination unit tests', () => {
     );
 
     expect(actual).toBeUndefined();
+  });
+
+  it('should return undefined when the destination is not valid and has an auth token expiration time', () => {
+    jest.useFakeTimers('modern');
+    const dummyJwt = { user_id: 'user', zid: 'tenant' };
+    const destination = {
+      ...destinationOne,
+      authTokens: [{ expiresIn: '60' } as DestinationAuthToken]
+    };
+    destinationCache.cacheRetrievedDestination(
+      dummyJwt,
+      destination,
+      IsolationStrategy.User
+    );
+    const retrieveDestination = () =>
+      destinationCache.retrieveDestinationFromCache(
+        dummyJwt,
+        destination.name!,
+        IsolationStrategy.User
+      );
+
+    expect(retrieveDestination()).toEqual(destination);
+
+    const minutesToExpire = 2;
+    jest.advanceTimersByTime(60000 * minutesToExpire);
+
+    expect(retrieveDestination()).toBeUndefined();
   });
 });